Ishan_Patel_project2/requirements.txt
psycopg[binary]
confluent-kafka==2.6.0
python-dotenv==1.0.1